4-sqlite-settings: move migrate and testdb to db package
