+++ internal/db/migrate/migrate_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	_ "github.com/mattn/go-sqlite3"
-	"github.com/willemschots/househunt/internal/migrate"
+	"github.com/willemschots/househunt/internal/db/migrate"
 )
 
 func Test_RunFS(t *testing.T) {
