+++ internal/db/migrate/testdata/error_in_migration/run_1/1_create_test_table.sql
