+++ internal/db/migrate/testdata/error_in_migration/run_2/2_insert_with_typo.sql
