+++ internal/db/migrate/testdata/removal_mismatch/run_1/1_create_test_table.sql
