+++ internal/db/migrate/testdata/removal_mismatch/run_2/1_create_test_table.sql
