+++ internal/db/migrate/testdata/removal_mismatch/run_2/2_add_row_to_test_table.sql
