+++ internal/db/migrate/testdata/rename_mismatch/run_1/1_create_test_table.sql
