+++ internal/db/migrate/testdata/rename_mismatch/run_2/1_create_test_table.sql
