+++ internal/db/testdb/testdb.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	_ "github.com/mattn/go-sqlite3"
-	"github.com/willemschots/househunt/internal/migrate"
+	"github.com/willemschots/househunt/internal/db/migrate"
 	"github.com/willemschots/househunt/migrations"
 )
